Shift order specification into the template params for converters. Added explicit tests for the dense/sparse converters.

+++ tests/src/_tests/simulate_vector.h
@@ -37,7 +37,7 @@ template<typename T>
 struct SparseDetails {
     std::vector<T> value;
     std::vector<int> index;
-    std::vector<long> ptr;
+    std::vector<size_t> ptr;
 };
 
 template<typename T>
